Implement Chronicle verification utility.
This lets you plug in a summary hash from a Chronicle and query a pool
of other Chronicle instances to ensure they have a record of it stored.

The envisioned use case:

1. Have a local-ish replica of the WordPress Chronicle.
2. For each update of the local-ish Chronicle, cross-reference against
   other instances to ensure some quorum (e.g. at least 3 of 5) of
   Chronicle instances see the same record.

This implements part 2 of that use case.

Part 1 (having a source Chronicle that is parsed to assemble Action
objects, which should allow anyone to deterministically reproduce the
current snapshot of trusted public keys and software updates) is the
next thing to be developed.

Continuing on:

3. Developer tools for signing releases.
4, Infrastructure plumbing for ensuring updates are streamed to the
   source Chronicle that seeds part 1.

And then we will develop other open source projects that consume
libgossamer in some which way.